WIP: add broken skip/widget
