+++ components/react/ModalDialog.tsx
@@ -9,7 +9,6 @@ import env from '../../config/env';
 export interface ModalDialogProps extends React.PropsWithChildren {
   open: boolean;
   onClose?: () => boolean | void;
-
   style?: React.CSSProperties;
   className?: string;
   panelClassName?: string;
