+++ global.d.ts
@@ -9,5 +9,6 @@ declare global {
 
   interface Window {
     keplr?: any;
+    ethereum?: any;
   }
 }
