+++ package.json
@@ -11,7 +11,7 @@
     "start": "next start",
     "lint": "next lint",
     "format": "prettier --write .",
-    "remove": "rm -rf node_modules/ bun.lockb $HOME/.bun/install/cache/",
+    "remove": "rm -rf node_modules/ bun.lockb $HOME/.bun/install/cache/ .next/",
     "update-deps": "bunx npm-check-updates --root --format group -i",
     "test:coverage": "bun test --coverage",
     "test:coverage:lcov": "bun run test:coverage --coverage-reporter=lcov --coverage-dir ./coverage",
@@ -51,6 +51,7 @@
     "@react-three/fiber": "^8.17.8",
     "@sinonjs/fake-timers": "^14.0.0",
     "@skip-go/client": "^0.16.16",
+    "@skip-go/widget": "^3.4.1",
     "@tailwindcss/postcss": "^4.0.10",
     "@tanstack/react-query": "^5.67.1",
     "@tanstack/react-query-devtools": "^5.67.1",
